+++ AppRoot/Controllers/online_courses_controller.php
@@ -37,29 +37,58 @@
     }
 
     function create($name, $description, $difficulty_level_id) {
-      $info = ["name" => $name, "description" => $description, "difficulty_level" => $difficulty_level_id];
-      $this->db->saveArray($this->tableName, $info);
+      if (isset($_SESSION["logged_user"]) && $_SESSION["logged_user"]["role"] == "admin") {
+        $info = ["name" => $name, "description" => $description, "difficulty_level" => $difficulty_level_id];
+        $this->db->saveArray($this->tableName, $info);
+        header('Location: ' . '../#/online-courses/' . $id . '/details');
+        exit();
+      } else {
+        header('Location: ' . '../#/home');
+        exit();
+      }
     }
 
     function edit($id, $name, $description, $difficulty_level_id) {
-      $info = ["id" => $id, "name" => $name, "description" => $description, "difficulty_level" => $difficulty_level_id];
-      $result = $this->db->saveArray($this->tableName, $info);
+      if (isset($_SESSION["logged_user"]) && $_SESSION["logged_user"]["role"] == "admin") {
+        $info = ["id" => $id, "name" => $name, "description" => $description, "difficulty_level" => $difficulty_level_id];
+        $result = $this->db->saveArray($this->tableName, $info);
+        header('Location: ' . '../#/online-courses/' . $id . '/details');
+        exit();
+      } else {
+        header('Location: ' . '../#/home');
+        exit();
+      }
     }
 
     function enrollStudent($courseId, $studentId) {
-      $name = "students_courses";
-      $info = ["student_id" => $studentId, "course_id" => $course_id];
-      $this->db->saveArray($name, $info);
+      if (isset($_SESSION["logged_user"])) {
+        $name = "students_courses";
+        $info = ["student_id" => $studentId, "course_id" => $course_id];
+        $this->db->saveArray($name, $info);
+      } else {
+        header('Location: ' . '../#/home');
+        exit();
+      }
     }
 
     function disenrollStudent($courseId, $studentId) {
-      $name = "students_courses";
-      $query = "DELETE FROM " . $name . " WHERE course_id = " . $courseId . " AND student_id = " . $student_id;
-      $result = $this->db->fetchArray($query);
+      if (isset($_SESSION["logged_user"])) {
+        $name = "students_courses";
+        $query = "DELETE FROM " . $name . " WHERE course_id = " . $courseId . " AND student_id = " . $student_id;
+        $result = $this->db->fetchArray($query);
+      } else {
+        header('Location: ' . '../#/home');
+        exit();
+      }
     }
 
     function delete($id) {
-      $result = $this->db->deleteRow($this->tableName, $id, "id");
+      if (isset($_SESSION["logged_user"]) && $_SESSION["logged_user"]["role"] == "admin") {
+        $result = $this->db->deleteRow($this->tableName, $id, "id");
+      } else {
+        header('Location: ' . '../#/home');
+        exit();
+      }
     }
 
     function getById($id) {
